Settings components moved to a settings module
