+++ packages/app/src/main/store/settings.interface.ts
@@ -3,10 +3,16 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import {
-  ConversionFormat,
-  Convert,
-} from "Renderer/components/rest/settings/audio-conversion-radio-group.enum"
+export enum Convert {
+  AlwaysAsk = "Always ask",
+  ConvertAutomatically = "Convert automatically",
+}
+
+export enum ConversionFormat {
+  FLAC = "FLAC",
+  WAV = "WAV",
+  MP3 = "MP3",
+}
 
 export interface SettingsUpdateOption {
   key: AppSettings[string]
