+++ packages/app/src/overview/components/overview-screens/pure-overview/pure-overview.component.test.tsx
@@ -10,8 +10,11 @@ import store from "Renderer/store"
 import history from "Renderer/routes/history"
 import { renderWithThemeAndIntl } from "Renderer/utils/render-with-theme-and-intl"
 import { PureOverview } from "App/overview/components/overview-screens/pure-overview/pure-overview.component"
-import { DataState, UpdatingState } from "Renderer/models/basic-info/basic-info.typings"
-import { ConversionFormat, Convert } from "Renderer/components/rest/settings/audio-conversion-radio-group.enum"
+import {
+  DataState,
+  UpdatingState,
+} from "Renderer/models/basic-info/basic-info.typings"
+import { ConversionFormat, Convert } from "App/main/store/settings.interface"
 import { StatusTestIds } from "App/overview/components/status/status-test-ids.enum"
 import { SystemTestIds } from "App/overview/components/system/system-test-ids.enum"
 import { intl } from "Renderer/utils/intl"
@@ -94,7 +97,7 @@ const defaultProps: ComponentProps<typeof PureOverview> = {
   memorySpace: {
     free: 100,
     full: 200,
-  }
+  },
 }
 
 const render = () => {
