--- packages/app/src/renderer/components/rest/settings/audio-conversion-radio-group.enum.ts
@@ -1,15 +1,0 @@
-/**
- * Copyright (c) Mudita sp. z o.o. All rights reserved.
- * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
- */
-
-export enum Convert {
-  AlwaysAsk = "Always ask",
-  ConvertAutomatically = "Convert automatically",
-}
-
-export enum ConversionFormat {
-  FLAC = "FLAC",
-  WAV = "WAV",
-  MP3 = "MP3",
-}
