+++ packages/app/src/renderer/modules/settings/components/settings-ui.component.tsx
@@ -17,7 +17,7 @@ import { noop } from "Renderer/utils/noop"
 import styled from "styled-components"
 import { FormattedMessage } from "react-intl"
 import { borderColor } from "Renderer/styles/theming/theme-getters"
-import SettingsToggler from "Renderer/components/rest/settings/settings-toggler.component"
+import SettingsToggler from "Renderer/modules/settings/components/settings-toggler.component"
 import { SettingsTestIds } from "Renderer/modules/settings/settings.enum"
 import Tooltip from "Renderer/components/core/tooltip/tooltip.component"
 import { defineMessages } from "react-intl"
