+++ packages/app/src/renderer/modules/settings/tabs/about/about-ui.component.tsx
@@ -15,7 +15,7 @@ import {
   SettingsLabel,
   SettingsTableRow,
   SettingsWrapper,
-} from "Renderer/components/rest/settings/settings-ui.component"
+} from "Renderer/modules/settings/components/settings-ui.component"
 import { AboutTestIds } from "Renderer/modules/settings/tabs/about/about.enum"
 import {
   letterSpacing,
