+++ packages/app/src/renderer/modules/settings/tabs/notifications/notifications.stories.tsx
@@ -5,7 +5,7 @@
 
 import { storiesOf } from "@storybook/react"
 import React from "react"
-import NotificationsUI from "Renderer/components/rest/settings/notifications-ui.component"
+import NotificationsUI from "Renderer/modules/settings/components/notifications-ui.component"
 
 storiesOf("Settings/Notifications", module).add("Notifications", () => (
   <div style={{ maxWidth: "63rem" }}>
